move video code to scripts.js
